Implement the remaining methods in DefaultPermissionProvider
